+++ src/Rebus/Timeout/IStoreTimeouts.cs
@@ -1,3 +1,4 @@
+using System;
 using System.Collections.Generic;
 
 namespace Rebus.Timeout
@@ -20,8 +21,43 @@ namespace Rebus.Timeout
 
         /// <summary>
         /// Should get all due timeouts. When a timeout has been properly processed, <see cref="DueTimeout.MarkAsProcessed"/>
-        /// should be called, which should cause the timeout to be removed from the underlying data store.
+        /// should be called, which should cause the timeout to be removed from the underlying data store. Also, please remember to
+        /// dispose the <see cref="DueTimeoutsResult"/> after the timeouts have been processed.
         /// </summary>
-        IEnumerable<DueTimeout> GetDueTimeouts();
+        DueTimeoutsResult GetDueTimeouts();
+    }
+
+    /// <summary>
+    /// Due timeout result wrapper that allows for returning a sequence of due timeouts along with an action to perform when the timeouts have been processed
+    /// </summary>
+    public class DueTimeoutsResult : IDisposable
+    {
+        readonly Action disposeAction;
+
+        /// <summary>
+        /// Creates the result wrapper with the given due timeout, optionally including an action to be performed when the result is disposed
+        /// </summary>
+        public DueTimeoutsResult(IEnumerable<DueTimeout> dueTimeouts, Action disposeAction = null)
+        {
+            this.disposeAction = disposeAction ?? (() => { });
+            DueTimeouts = dueTimeouts;
+        }
+
+
+        /// <summary>
+        /// Gets the sequence of due timeout that were returned
+        /// </summary>
+        public IEnumerable<DueTimeout> DueTimeouts { get; private set; }
+        
+        public void Dispose()
+        {
+            Dispose(true);
+            GC.SuppressFinalize(this);
+        }
+
+        public void Dispose(bool disposing)
+        {
+            disposeAction();
+        }
     }
 }